+++ end-to-end-tests/src/test/java/fi/jumi/test/BuildTest.java
@@ -4,6 +4,8 @@
 
 package fi.jumi.test;
 
+import com.google.common.base.Predicate;
+import com.google.common.collect.Iterables;
 import fi.luontola.buildtest.*;
 import org.hamcrest.Matchers;
 import org.junit.Test;
@@ -156,7 +158,16 @@ public class BuildTest {
 
     @Test
     public void deprecated_methods_are_removed_after_the_transition_period() throws IOException {
-        expectedDeprecations.verify(new ClassesInJarFile(getProjectJar()));
+        expectedDeprecations.verify(ignoreShadedLibraries(new ClassesInJarFile(getProjectJar())));
+    }
+
+    private static Iterable<ClassNode> ignoreShadedLibraries(Iterable<ClassNode> classes) {
+        return Iterables.filter(classes, new Predicate<ClassNode>() {
+            @Override
+            public boolean apply(ClassNode c) {
+                return !c.name.contains("/INTERNAL/");
+            }
+        });
     }